+++ tests/chapter09/AbstractClass.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+/**
+ * Class AbstractClass
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+abstract class AbstractClass
+{
+    public function concreteMethod()
+    {
+        return $this->abstractMethod();
+    }
+
+    public abstract function abstractMethod();
+}
+