+++ tests/chapter09/AbstractTrait.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+/**
+ * Class AbstractTrait
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+trait AbstractTrait
+{
+    public function concreteMethod()
+    {
+        return $this->abstractMethod();
+    }
+
+    public abstract function abstractMethod();
+}