+++ tests/chapter09/Observer.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+/**
+ * Class Observer
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+class Observer
+{
+    public function update($argument)
+    {
+        // Do something.
+    }
+
+    public function reportError($errorCode, $errorMessage, Subject $subject)
+    {
+        // Do something
+    }
+
+    // Other methods.
+}