+++ tests/chapter09/Stub4Test.php
@@ -1,0 +1,39 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class Stub4Test
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+class Stub4Test extends TestCase
+{
+    public function testReturnArgumentStub()
+    {
+        // Create a stub for the SomeClass class.
+        $stub = $this->createMock(SomeClass::class);
+
+        // Configure the stub.
+        $stub->method('doSomething')
+            ->will($this->returnArgument(0));
+
+        // $stub->doSomething('foo') returns 'foo'
+        $this->assertEquals('foo', $stub->doSomething('foo'));
+
+        // $stub->doSomething('bar') returns 'bar'
+        $this->assertEquals('bar', $stub->doSomething('bar'));
+    }
+}