+++ tests/chapter09/Stub7Test.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class Stub7Test
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+class Stub7Test extends TestCase
+{
+    public function testReturnCallbackStub()
+    {
+        // Create a stub for the SomeClass class.
+        $stub = $this->createMock(SomeClass::class);
+
+        // Configure the stub.
+        $stub->method('doSomething')
+            ->will($this->returnCallback('str_rot13'));
+
+        // $stub->doSomething($argument) returns str_rot13($argument)
+        $this->assertEquals('fbzrguvat', $stub->doSomething('something'));
+    }
+}