+++ tests/chapter09/Stub8Test.php
@@ -1,0 +1,38 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class Stub8Test
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+class Stub8Test extends TestCase
+{
+    public function testOnConsecutiveCallsStub()
+    {
+        // Create a stub for the SomeClass class.
+        $stub = $this->createMock(SomeClass::class);
+
+        // Configure the stub.
+        $stub->method('doSomething')
+            ->will($this->onConsecutiveCalls(2, 3, 5, 7));
+
+        // $stub->doSomething() returns a different value each time
+        $this->assertEquals(2, $stub->doSomething());
+        $this->assertEquals(3, $stub->doSomething());
+        $this->assertEquals(5, $stub->doSomething());
+    }
+}