+++ tests/chapter09/Stub9Test.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class Stub9Test
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+class Stub9Test extends TestCase
+{
+    public function testThrowExceptionStub()
+    {
+        // Create a stub for the SomeClass class.
+        $stub = $this->createMock(SomeClass::class);
+
+        // Configure the stub.
+        $stub->method('doSomething')
+            ->will($this->throwException(new \Exception));
+
+        // $stub->doSomething() throws Exception
+        $stub->doSomething();
+    }
+}