+++ tests/chapter09/Subject4Test.php
@@ -1,0 +1,45 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class Subject4Test
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+class Subject4Test extends TestCase
+{
+    public function testObserversAreUpdated()
+    {
+        $subject = new Subject('My subject');
+
+        // Create a prophecy for the Observer class.
+        $observer = $this->prophesize(Observer::class);
+
+        // Set up the expectation for the update() method
+        // to be called only once and with the string 'something'
+        // as its parameter.
+        $observer->update('something')->shouldBeCalled();
+
+        // Reveal the prophecy and attach the mock object
+        // to the Subject.
+        $subject->attach($observer->reveal());
+
+        // Call the doSomething() method on the $subject object
+        // which we expect to call the mocked Observer object's
+        // update() method with the string 'something'.
+        $subject->doSomething();
+    }
+}