+++ tests/chapter09/SubjectTest.php
@@ -1,0 +1,49 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter09;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class SubjectTest
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter09
+ */
+class SubjectTest extends TestCase
+{
+    public function testObserversAreUpdated()
+    {
+        // Create a mock for the Observer class,
+        // only mock the update() method.
+        $observer = $this->getMockBuilder(Observer::class)
+            ->setMethods(['update'])
+            ->getMock();
+
+        // Set up the expectation for the update() method
+        // to be called only once and with the string 'something'
+        // as its parameter.
+        $observer->expects($this->once())
+            ->method('update')
+            ->with($this->equalTo('something'));
+
+        // Create a Subject object and attach the mocked
+        // Observer object to it.
+        $subject = new Subject('My subject');
+        $subject->attach($observer);
+
+        // Call the doSomething() method on the $subject object
+        // which we expect to call the mocked Observer object's
+        // update() method with the string 'something'.
+        $subject->doSomething();
+    }
+}